Add kuttl tests to check keystone config changes
Add kuttl test to check that changing the keystone service config
triggers a new pod with the new configuration. The tests check that
logging by default does not include debug information, then changes
the keystone config to set debug = true, and then checks that the logs
of the new pod now include debug information. As a last step, it checks
that the kesytone service is cleaned up.